requirements.txt
flask
gunicorn
requests
python-dotenv
ibm-watson
ibm-cloud-sdk-core